requirements.txt
Django==3.1.1
django-mysql==3.8.1
django-cuser==2017.3.16
watchdog==0.10.3
xlrd==1.2.0
XlsxWriter==1.3.3
pandas==1.1.2
djangorestframework==3.11.1
django-cors-headers==3.5.0
requests==2.24.0
coverage==5.2.1
pylint==2.6.0
python-dateutil==2.8.1
Pympler==0.8
sqlparse==0.3.1
mysqlclient==1.4.6
httpx==0.14.3